+++ src/dgisim/encoding/mappings.py
@@ -223,6 +223,7 @@ CARD_MAPPING: dict[type["Card"], int] = HashableDict({
         # 158 for enhanced Flowing Rings
         (card.HeartOfKhvarenasBrilliance, 159),
         (card.VourukashasGlow, 160),
+        (card.EchoesOfAnOffering, 161),
 
         #### 2 Support Card 0250 ####
         #### 2.1 Location Card 0250 ####
@@ -649,6 +650,7 @@ STT_MAPPING: dict[type["Status"], int] = HashableDict({
         # 208 for enhanced Flowing Rings
         (status.HeartOfKhvarenasBrillianceStatus, 209),
         (status.VourukashasGlowStatus, 210),
+        (status.EchoesOfAnOfferingStatus, 211),
 
         #### 2 Combat Status 300 ####
         #### 2.1 General Combat Status 300 ####